+++ src/routes/development/+page.svelte
@@ -179,6 +179,11 @@
         stringify: compileJSONPointer
       },
       label: 'JSONPointer'
+    },
+    {
+      id: 'JSON',
+      value: JSON,
+      label: 'JSON'
     }
   ]